requirements.txt
python-docx
html2text